requirements.txt
flask==3.0.3
gunicorn==22.0.0
google-cloud-aiplatform>=1.70.0
python-dotenv==1.0.0
requests>=2.32.3
oauthlib==3.2.2
requests-oauthlib==1.3.1
